refactor(router): extract createSpaRouter factory
Seven SPA routers (CAHFS, CMS, CTS, ClinicalScheduler, Computing, Effort,
Students) duplicated the same createRouter + VITE_VIPER_HOME + scroll-top
+ useRouteFocus setup. Hoist into src/shared/createSpaRouter.ts so each
router only owns its bespoke beforeEach guard. No behavior change.
