+++ VueApp/src/CTS/router/index.ts
@@ -1,15 +1,9 @@
-import { createRouter, createWebHistory } from "vue-router"
+import { createSpaRouter } from "@/shared/createSpaRouter"
 import { routes } from "./routes"
 import { useRequireLogin } from "@/composables/RequireLogin"
 import { checkHasOnePermission } from "@/composables/CheckPagePermission"
-import { useRouteFocus } from "@/composables/use-route-focus"
 
-const baseUrl = import.meta.env.VITE_VIPER_HOME
-const router = createRouter({
-    scrollBehavior: () => ({ left: 0, top: 0 }),
-    history: createWebHistory(baseUrl),
-    routes,
-})
+const router = createSpaRouter(routes)
 
 router.beforeEach(async (to) => {
     const { requireLogin } = useRequireLogin(to)
@@ -24,7 +18,5 @@ router.beforeEach(async (to) => {
         }
     }
 })
-
-useRouteFocus(router)
 
 export { router }
