+++ VueApp/src/ClinicalScheduler/router/index.ts
@@ -1,20 +1,12 @@
-import { createRouter, createWebHistory } from "vue-router"
+import { createSpaRouter } from "@/shared/createSpaRouter"
 import { clinicalSchedulerRoutes as routes } from "./routes"
 import { useRequireLogin } from "@/composables/RequireLogin"
-import { useRouteFocus } from "@/composables/use-route-focus"
 
-const baseUrl = import.meta.env.VITE_VIPER_HOME
-const router = createRouter({
-    scrollBehavior: () => ({ left: 0, top: 0 }),
-    history: createWebHistory(baseUrl),
-    routes,
-})
+const router = createSpaRouter(routes)
 
 router.beforeEach((to) => {
     const { requireLogin } = useRequireLogin(to)
     return requireLogin(true, "SVMSecure.ClnSched")
 })
-
-useRouteFocus(router)
 
 export { router as clinicalSchedulerRouter }
